+++ app/controllers/concerns/shopify_app/require_known_shop.rb
@@ -7,9 +7,11 @@ module ShopifyApp
 
     included do
       if ancestors.include?(ShopifyApp::LoginProtection)
-        raise ConfigurationError,
-          "You are attempting to use both RequireKnownShop and LoginProtection in the same controller. "\
-            "These are not compatible."
+        ActiveSupport::Deprecation.warn(<<~EOS)
+          We detected the use of incompatible concerns (RequireKnownShop and LoginProtection) in #{name},
+          which may lead to unpredictable behavior. In a future release of this library this will raise an error.
+        EOS
+
       end
 
       before_action :check_shop_domain