+++ test/controllers/concerns/require_known_shop_test.rb
@@ -62,17 +62,27 @@ class RequireKnownShopTest < ActionController::TestCase
   end
 
   test "detects incompatible controller concerns" do
-    assert_raises ShopifyApp::ConfigurationError do
+    assert_deprecated(/incompatible concerns/) do
       Class.new(ApplicationController) do
         include ShopifyApp::RequireKnownShop
         include ShopifyApp::LoginProtection
       end
     end
 
-    assert_raises ShopifyApp::ConfigurationError do
+    assert_deprecated(/incompatible concerns/) do
       Class.new(ApplicationController) do
         include ShopifyApp::RequireKnownShop
         include ShopifyApp::Authenticated # since this indirectly includes LoginProtection
+      end
+    end
+
+    assert_deprecated(/incompatible concerns/) do
+      authenticated_controller = Class.new(ApplicationController) do
+        include ShopifyApp::Authenticated
+      end
+
+      Class.new(authenticated_controller) do
+        include ShopifyApp::RequireKnownShop
       end
     end
   end